refactor: split `types.ts` into multi `types/*` files, update imports
