+++ src/BatchRequests/index.ts
@@ -1,3 +1,4 @@
 export * from "./batchRequestWithExponentialBackoff.js";
 export * from "./handleBatchRequests.js";
-export * from "./types.js";
+
+export * from "./types/index.js";
