+++ src/BatchRequests/types/index.ts
@@ -1,0 +1,3 @@
+export * from "./BatchOperationParams.js";
+export * from "./BatchRetryExponentialBackoffConfigs.js";
+export * from "./BatchRequestFunction.js";
